Add support for addColumn for typed table
