+++ tests/Backend/Synapse/TableDefinitionOperationsTest.php
@@ -2,6 +2,8 @@
 
 namespace Keboola\Test\Backend\Synapse;
 
+use Keboola\Csv\CsvFile;
+use Keboola\StorageApi\Metadata;
 use Keboola\Test\StorageApiTestCase;
 
 class TableDefinitionOperationsTest extends StorageApiTestCase
@@ -57,10 +59,249 @@ class TableDefinitionOperationsTest extends StorageApiTestCase
         return $this->_client->createTableDefinition($bucketId, $data);
     }
 
+    public function testDataPreviewForTableDefinitionWithDecimalType()
+    {
+        $bucketId = $this->getTestBucketId(self::STAGE_IN);
+
+        $tableDefinition = [
+            'name' => 'my-new-table-data-preview',
+            'primaryKeysNames' => ['id'],
+            'columns' => [
+                [
+                    'name' => 'id',
+                    'definition' => [
+                        'type' => 'INT',
+                    ],
+                ],
+                [
+                    'name' => 'column_decimal',
+                    'definition' => [
+                        'type' => 'DECIMAL',
+                        'length' => '4,3'
+                    ],
+                ],
+                [
+                    'name' => 'column_float',
+                    'definition' => [
+                        'type' => 'FLOAT',
+                    ],
+                ],
+                [
+                    'name' => 'column_boolean',
+                    'definition' => [
+                        'type' => 'BIT',
+                    ],
+                ],
+                [
+                    'name' => 'column_date',
+                    'definition' => [
+                        'type' => 'DATE',
+                    ],
+                ],
+                [
+                    'name' => 'column_timestamp',
+                    'definition' => [
+                        'type' => 'TIME',
+                    ],
+                ],
+                [
+                    'name' => 'column_varchar',
+                    'definition' => [
+                        'type' => 'VARCHAR',
+                    ],
+                ],
+                [
+                    'name' => 'column_money',
+                    'definition' => [
+                        'type' => 'MONEY',
+                    ],
+                ],
+                [
+                    'name' => 'column_small_money',
+                    'definition' => [
+                        'type' => 'SMALLMONEY',
+                    ],
+                ],
+                [
+                    'name' => 'column_uniq',
+                    'definition' => [
+                        'type' => 'UNIQUEIDENTIFIER',
+                    ],
+                ],
+            ],
+            'distribution' => [
+                'type' => 'HASH',
+                'distributionColumnsNames' => ['id'],
+            ],
+        ];
+
+        $csvFile = new CsvFile(tempnam(sys_get_temp_dir(), 'keboola'));
+        $csvFile->writeRow([
+            'id',
+            'column_decimal',
+            'column_float',
+            'column_boolean',
+            'column_date',
+            'column_timestamp',
+            'column_varchar',
+            'column_money',
+            'column_small_money',
+            'column_uniq',
+        ]);
+        $csvFile->writeRow(['1', '003.123', '3.14', 1, '1989-08-31', '05:00:01', 'roman', '3148.29', '3148.29', '0E984725-C51C-4BF4-9960-E1C80E27ABA0']);
+
+
+        $tableId = $this->_client->createTableDefinition($bucketId, $tableDefinition);
+
+        $this->_client->writeTable($tableId, $csvFile);
+
+        $data = $this->_client->getTableDataPreview($tableId, ['format' => 'json']);
+
+        $expectedPreview = [
+            [
+                [
+                    'columnName' => 'id',
+                    'value' => '1',
+                    'isTruncated' => false,
+                ],
+                [
+                    'columnName' => 'column_decimal',
+                    'value' => '3.123',
+                    'isTruncated' => false,
+                ],
+                [
+                    'columnName' => 'column_float',
+                    'value' => '3.14',
+                    'isTruncated' => false,
+                ],
+                [
+                    'columnName' => 'column_boolean',
+                    'value' => '1',
+                    'isTruncated' => false,
+                ],
+                [
+                    'columnName' => 'column_date',
+                    'value' => '1989-08-31',
+                    'isTruncated' => false,
+                ],
+                [
+                    'columnName' => 'column_timestamp',
+                    'value' => '05:00:01.0000000',
+                    'isTruncated' => false,
+                ],
+                [
+                    'columnName' => 'column_varchar',
+                    'value' => 'roman',
+                    'isTruncated' => false,
+                ],
+                [
+                    'columnName' => 'column_money',
+                    'value' => '3148.29',
+                    'isTruncated' => false,
+                ],
+                [
+                    'columnName' => 'column_small_money',
+                    'value' => '3148.29',
+                    'isTruncated' => false,
+                ],
+                [
+                    'columnName' => 'column_uniq',
+                    'value' => '0E984725-C51C-4BF4-9960-E1C80E27ABA0',
+                    'isTruncated' => false,
+                ],
+            ],
+        ];
+
+        $this->assertSame(
+            $expectedPreview,
+            $data['rows']
+        );
+
+        $this->assertSame(1, count($data['rows']));
+
+        //test types is provided from source table for alias
+        $firstAliasTableId = $this->_client->createAliasTable($this->getTestBucketId(self::STAGE_IN), $tableId, 'table-1');
+
+        $data = $this->_client->getTableDataPreview($firstAliasTableId, ['format' => 'json']);
+
+        $this->assertSame(
+            $expectedPreview,
+            $data['rows']
+        );
+
+        $this->assertSame(1, count($data['rows']));
+    }
+
     public function testAddColumnOnTypedTable()
     {
-        $this->expectExceptionMessage("Not implemented for typed tables");
-        $this->_client->addTableColumn($this->tableId, 'newColumn');
+        $tableDefinition = [
+            'name' => 'my-new-table-add-column',
+            'primaryKeysNames' => ['id'],
+            'columns' => [
+                [
+                    'name' => 'id',
+                    'definition' => [
+                        'type' => 'INT',
+                    ],
+                ],
+                [
+                    'name' => 'column_decimal',
+                    'definition' => [
+                        'type' => 'DECIMAL',
+                        'length' => '4,3'
+                    ],
+                ],
+            ],
+            'distribution' => [
+                'type' => 'HASH',
+                'distributionColumnsNames' => ['id'],
+            ],
+        ];
+
+        $sourceTableId = $this->_client->createTableDefinition($this->getTestBucketId(self::STAGE_IN), $tableDefinition);
+
+        $firstAliasTableId = $this->_client->createAliasTable($this->getTestBucketId(self::STAGE_IN), $sourceTableId, 'table-1');
+        $secondAliasTableId = $this->_client->createAliasTable($this->getTestBucketId(self::STAGE_IN), $firstAliasTableId, 'table-2');
+
+        $this->_client->addTableColumn($sourceTableId, 'added_column', [
+            'type' => 'DECIMAL',
+            'length' => '4,3'
+        ]);
+
+        $expectedColumns = ['id', 'column_decimal', 'added_column'];
+        $this->assertEquals($expectedColumns, $this->_client->getTable($sourceTableId)['columns']);
+        $this->assertEquals($expectedColumns, $this->_client->getTable($firstAliasTableId)['columns']);
+        $this->assertEquals($expectedColumns, $this->_client->getTable($secondAliasTableId)['columns']);
+
+        // check that the new table has correct datypes in metadata
+        $metadataClient = new Metadata($this->_client);
+        $addedColumnMetadata = $metadataClient->listColumnMetadata("{$sourceTableId}.added_column");
+        // alias tables has metadata from source table
+        $firstAliasAddedColumnMetadata = $this->_client->getTable($firstAliasTableId)['sourceTable']['columnMetadata']['added_column'];
+        $secondAliasAddedColumnMetadata = $this->_client->getTable($secondAliasTableId)['sourceTable']['columnMetadata']['added_column'];
+
+        foreach ([$addedColumnMetadata, $firstAliasAddedColumnMetadata, $secondAliasAddedColumnMetadata] as $columnMetadata) {
+            $this->assertArrayEqualsExceptKeys([
+                'key' => 'KBC.datatype.type',
+                'value' => 'DECIMAL',
+                'provider' => 'storage',
+            ], $columnMetadata[0], ['id', 'timestamp']);
+            $this->assertArrayEqualsExceptKeys([
+                'key' => 'KBC.datatype.nullable',
+                'value' => '1',
+                'provider' => 'storage',
+            ], $columnMetadata[1], ['id', 'timestamp']);
+            $this->assertArrayEqualsExceptKeys([
+                'key' => 'KBC.datatype.basetype',
+                'value' => 'NUMERIC',
+                'provider' => 'storage',
+            ], $columnMetadata[2], ['id', 'timestamp']);
+            $this->assertArrayEqualsExceptKeys([
+                'key' => 'KBC.datatype.length',
+                'value' => '4,3',
+                'provider' => 'storage',
+            ], $columnMetadata[3], ['id', 'timestamp']);
+        }
     }
 
     public function testDropColumnOnTypedTable()
